feat: add `SetupProperty<TProperty>(...)` method for property stubbing to `Mimic<T>`
Adds a new setup method `SetupProperty<TProperty>(Expression<Func<T, TProperty>> propertyExpression, TProperty? initialValue = default)` on `Mimic<T>` that can be used to enable property stubbing for the supplied property.

Property stubbing allows you to enable a backing field for the property on the mimicked interface so that calls to get return any previously set value or the default value.