feat: Implement `Strict` mode (default: `true`) to require an invocation to have a matching setup or else throw an exception
Implements `Strict` mode, set using the property of the same name on `Mimic<T>` which is `true` by default, the same as existing functionality.

When set to `true` a matching setup is mandatory for an invocation and without one a `MimicException` will be thrown to indicate a required setup is missing.

When set to `false` we are more lenient. And where there is no matching setup and the method returns a non-void value, we will attempt to get the default value of the return type and return that instead. There are also a few special cases for; Arrays, IEnumerables and Tuples.