feat: Add `Verifiable` flag to setups and `Verify` method on `Mimic<T>`
Adds a new `Verifiable` flag on setups that can be set for a particular method setup in order to verify that it has been called at the end of a test.

Execution of the verifiable setup can be checked by calling `Verify` on the corresponding `Mimic<T>` the setup belongs to.